added toast messages, renamed openapi folder
