+++ frontend/package.json
@@ -45,6 +45,7 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-hook-form": "^7.51.0",
+    "react-hot-toast": "^2.4.1",
     "sharp": "^0.33.2",
     "zod": "^3.22.4"
   },
