+++ frontend/src/app/layout.tsx
@@ -1,11 +1,16 @@
 import { UserProvider } from '@auth0/nextjs-auth0/client';
 import type { Metadata } from "next";
 import "./globals.css";
+import { Toaster } from "react-hot-toast";
 
 export const metadata: Metadata = {
   title: "Compass",
   description: "",
 };
+
+const toastOptions = {
+  duration: 4000
+}
 
 export default function RootLayout({
   children,
@@ -15,6 +20,7 @@ export default function RootLayout({
   return (
     <html lang="en">
       <body>
+        <Toaster position="bottom-center" toastOptions={toastOptions}/>
         <UserProvider>{children}</UserProvider>
       </body>
     </html>
