+++ frontend/src/constants/toastMessages.tsx
@@ -1,0 +1,9 @@
+enum toastMessages {
+  USERS_NOT_LOADED = "User konnten nicht geladen werden",
+  USER_CREATED = "User erstellt",
+  USER_NOT_CREATED = "User konnte nicht erstellt werden",
+  USER_UPDATED = "User aktualisiert",
+  USER_NOT_UPDATED = "User konnte nicht aktualisiert werden",
+}
+
+export default toastMessages;
