+++ frontend/src/openapi/compassClient/.npmignore
