+++ frontend/src/openapi/compassClient/.openapi-generator-ignore
