+++ frontend/src/openapi/compassClient/.openapi-generator/VERSION
