+++ frontend/src/openapi/compassClient/base.ts
