+++ frontend/src/openapi/compassClient/configuration.ts
