+++ frontend/src/openapi/compassClient/package.json
